# packages/daemon/openapi.yaml
openapi: 3.0.3
info:
  title: orch8 API
  description: |
    REST + WebSocket API for the orch8 orchestration platform.

    ## Authentication

    Two authentication paths are supported on all endpoints (except `GET /api/health`):

    **Agent Authentication** — Provide both `X-Agent-Id` and `X-Project-Id` headers. The agent is validated against the database.

    **Admin Authentication** — Requests from localhost (`127.0.0.1`, `::1`) without agent headers are treated as admin. Admins bypass all permission checks.
  version: 0.0.0
  license:
    name: Private
servers:
  - url: http://localhost:3847
    description: Local daemon
...
components:
  securitySchemes:
    AgentAuth:
      type: apiKey
      in: header
      name: X-Agent-Id
      description: Agent ID header (must be paired with X-Project-Id)
    ProjectScope:
      type: apiKey
      in: header
      name: X-Project-Id
      description: Project ID scope header
    RunContext:
      type: apiKey
      in: header
      name: X-Run-Id
      description: Optional run ID for correlation
  schemas:
    Error:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          example: not_found
        message:
          type: string
          example: Task not found
    ValidationError:
      type: object
      required: [error, details]
      properties:
        error:
          type: string
          enum: [validation_error]
        details:
          type: array
          items:
            type: object
            properties:
              code:
                type: string
              path:
                type: array
                items:
                  type: string
              message:
                type: string
...
    Project:
      type: object
      properties:
        id:
          type: string
          example: proj_abc123
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        homeDir:
          type: string
        worktreeDir:
          type: string
        repoUrl:
          type: string
          nullable: true
        defaultBranch:
          type: string
        defaultModel:
          type: string
          nullable: true
        defaultMaxTurns:
          type: integer
          nullable: true
        budgetLimitUsd:
          type: number
          nullable: true
        budgetSpentUsd:
          type: number
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateProject:
      type: object
      required: [name, slug, homeDir, worktreeDir]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        slug:
          type: string
          minLength: 1
          maxLength: 100
          pattern: ^[a-z0-9-]+$
        description:
          type: string
          maxLength: 5000
          default: ""
        homeDir:
          type: string
          minLength: 1
        worktreeDir:
          type: string
          minLength: 1
        repoUrl:
          type: string
        defaultBranch:
          type: string
          default: main
        defaultModel:
          type: string
        defaultMaxTurns:
          type: integer
          minimum: 1
        budgetLimitUsd:
          type: number
          minimum: 0
    UpdateProject:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          maxLength: 5000
        repoUrl:
          type: string
          nullable: true
        defaultBranch:
          type: string
        defaultModel:
          type: string
          nullable: true
        defaultMaxTurns:
          type: integer
          minimum: 1
          nullable: true
        budgetLimitUsd:
          type: number
          minimum: 0
          nullable: true
        active:
          type: boolean

    # --- Task ---
...
    Run:
      type: object
      properties:
        id:
          type: string
          example: run_abc123
        agentId:
          type: string
        projectId:
          type: string
        taskId:
          type: string
          nullable: true
        invocationSource:
          $ref: "#/components/schemas/WakeupSource"
        triggerDetail:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/RunStatus"
        startedAt:
          type: string
          format: date-time
          nullable: true
        finishedAt:
          type: string
          format: date-time
          nullable: true
        processPid:
          type: integer
          nullable: true
        exitCode:
          type: integer
          nullable: true
        error:
          type: string
          nullable: true
        costUsd:
          type: number
          nullable: true
        model:
          type: string
          nullable: true
        sessionIdBefore:
          type: string
          nullable: true
        sessionIdAfter:
          type: string
          nullable: true
        logStore:
          type: string
          nullable: true
        logRef:
          type: string
          nullable: true
        parentRunId:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time

    # --- Knowledge ---
...
    RunEvent:
      type: object
      properties:
        id:
          type: string
          example: evt_abc123
        runId:
          type: string
        projectId:
          type: string
        seq:
          type: integer
        timestamp:
          type: string
          format: date-time
        eventType:
          type: string
        toolName:
          type: string
          nullable: true
        summary:
          type: string
        payload:
          type: object
          nullable: true
        createdAt:
          type: string
          format: date-time

    # --- Enums ---
...
    RunStatus:
      type: string
      enum: [queued, running, succeeded, failed, timed_out, cancelled]
    WakeupSource:
      type: string
      enum: [timer, assignment, on_demand, automation]
...
paths:
...
  /api/projects:
    post:
      tags: [Projects]
      summary: Create project
      description: Admin only.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProject"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
        "403":
          description: Not admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Slug already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    get:
      tags: [Projects]
      summary: List projects
      parameters:
        - name: active
          in: query
          schema:
            type: boolean
          description: Filter by active status
      responses:
        "200":
          description: Project list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
  /api/projects/{id}:
    get:
      tags: [Projects]
      summary: Get project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      tags: [Projects]
      summary: Update project
      description: Admin only.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProject"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
        "403":
          description: Not admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/projects/{id}/archive:
    post:
      tags: [Projects]
      summary: Archive project
      description: Admin only. Sets active to false.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Archived
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "403":
          description: Not admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ============================================================
  # Tasks
  # ============================================================
...
  /api/runs:
    get:
      tags: [Runs]
      summary: List runs
      description: Agent callers must have a projectId. Admins may omit for cross-project listing.
      parameters:
        - name: agentId
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/RunStatus"
        - name: taskId
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 500
            default: 100
      responses:
        "200":
          description: Run list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Run"
        "400":
          description: Agent without projectId
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/runs/{id}:
    get:
      tags: [Runs]
      summary: Get run
      description: Requires projectId.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Run
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Run"
        "400":
          description: Missing projectId
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/runs/{id}/cancel:
    post:
      tags: [Runs]
      summary: Cancel run
      description: Cancels a queued or running run. Requires projectId.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Run"
        "400":
          description: Missing projectId
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Run not in cancellable state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/runs/{id}/retry:
    post:
      tags: [Runs]
      summary: Retry a failed or cancelled run
      description: |
        Re-dispatches a failed or cancelled run for the same agent and task
        using the heartbeat wakeup machinery (so budget checks, pause state,
        and queue coalescing all apply). Requires projectId.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "202":
          description: Retry wakeup enqueued
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  wakeupId:
                    type: string
                  newRunId:
                    type: string
                    nullable: true
        "400":
          description: Missing projectId
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Run or agent not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Run not in a retryable state or agent is paused
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/runs/{id}/events:
    get:
      tags: [Runs]
      summary: Get structured events for a run
      description: Returns events ordered by sequence number. Requires projectId.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Event list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RunEvent"
        "400":
          description: Missing projectId
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/runs/{id}/log:
    get:
      tags: [Runs]
      summary: Get run log
      description: Requires projectId. Returns the full log content unless `tail` is provided.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: tail
          in: query
          schema:
            type: integer
            minimum: 1
          description: If provided, returns only the last N non-empty lines.
      responses:
        "200":
          description: Log content
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: string
                    description: Log file content (optionally tailed)
                  store:
                    type: string
                    nullable: true
                    description: Log storage backend identifier
                  bytes:
                    type: integer
                    description: Total byte length of the full log file
        "400":
          description: Missing projectId
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found or no log available
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ============================================================
  # Memory
  # ============================================================
...